Revert "Set up bundling for web ext code"
This reverts commit 20d23fac4fa52a0d469cb946a41feefb15edc3b8.
