--- cycle1/.babelrc
@@ -1,5 +1,0 @@
-{
-  "presets": [
-        "@babel/preset-env",
-    ]
-}
