+++ cycle1/.gitignore
@@ -1,6 +1,2 @@
 main.js
 wallet-popup.js
-
-node_modules
-
-*-bundle.js
