--- cycle1/.prettierrc.js
@@ -1,7 +1,0 @@
-module.exports = {
-  semicolon: false,
-  singleQuote: true,
-  tabWidth: 2,
-  tabs: false,
-  trailingComma: "es5"
-}
