--- cycle1/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "cycle1",
-  "version": "1.0.0",
-  "description": "The goal of the first prototype is to focus on how OSS projects interested in oscoin will be able to explore and interact with the network. We are limiting the scope to one type of user, meaning we're looking at the different intents that user type might have and how they then will interact with the network. We will have time exploring other types of users and advanced interactions in later iterations.",
-  "main": ".eslintrc.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "@babel/core": "^7.4.4",
-    "@babel/preset-env": "^7.4.4",
-    "babel-core": "^6.26.3",
-    "babel-loader": "^8.0.5",
-    "babelify": "^10.0.0",
-    "browserify": "^16.2.3",
-    "eslint": "5.16.0",
-    "gulp": "^4.0.1",
-    "vinyl-source-stream": "^2.0.0"
-  },
-  "dependencies": {
-    "base32": "^0.0.6",
-    "blakejs": "^1.1.0",
-    "tweetnacl": "^1.0.1"
-  }
-}
