+++ cycle1/public/manifest.json
@@ -5,10 +5,13 @@
   "version": "1.0",
   "homepage_url": "https://oscoin.io",
   "icons": {
-    "48": "icons/wallet-default.svg",
-    "96": "icons/wallet-default.svg"
+    "48": "icons/wallet.svg",
+    "96": "icons/wallet.svg"
   },
-  "permissions": ["activeTab", "*://localhost:8000/*"],
+  "permissions": [
+    "activeTab",
+    "*://localhost:8000/*"
+  ],
   "browser_action": {
     "default_popup": "wallet-popup.html#keys",
     "default_title": "oscoin wallet"
@@ -19,8 +22,14 @@
   },
   "content_scripts": [
     {
-      "matches": ["file://*/*", "http://*/*", "https://*/*"],
-      "js": ["scripts/wallet-contentscript.js"],
+      "matches": [
+        "file://*/*",
+        "http://*/*",
+        "https://*/*"
+      ],
+      "js": [
+        "scripts/wallet-contentscript.js"
+      ],
       "run_at": "document_start",
       "all_frames": true
     }
