+++ cycle1/public/wallet-background.html
@@ -2,6 +2,8 @@
 <html>
   <head>
     <meta charset="utf-8">
-    <script src="scripts/wallet-background-bundle.js"></script>
+    <script type="module" src="scripts/base32.js"></script>
+    <script src="scripts/nacl-fast-min.js"></script>
+    <script type="module" src="scripts/wallet-background.js"></script>
   </head>
 </html>
